+++ benches/generators.rs
@@ -1,19 +1,24 @@
+use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use bulletproofs::{BulletproofGens, PedersenGens};
-
-#[macro_use]
-extern crate criterion;
-use criterion::Criterion;
 
 fn pc_gens(c: &mut Criterion) {
     c.bench_function("PedersenGens::new", |b| b.iter(|| PedersenGens::default()));
 }
 
 fn bp_gens(c: &mut Criterion) {
-    c.bench_function_over_inputs(
-        "BulletproofGens::new",
-        |b, size| b.iter(|| BulletproofGens::new(*size, 1)),
-        (0..10).map(|i| 2 << i),
-    );
+    let mut group = c.benchmark_group("BulletproofGens::new");
+
+    for size in (0..10).map(|i| 2 << i) {
+        group.bench_with_input(
+            BenchmarkId::from_parameter(size),
+            &size,
+            |b, &s| {
+                b.iter(|| BulletproofGens::new(s, 1))
+            },
+        );
+    }
+
+    group.finish();
 }
 
 criterion_group! {