+++ benches/range_proof.rs
@@ -1,7 +1,4 @@
-#![allow(non_snake_case)]
-#[macro_use]
-extern crate criterion;
-use criterion::Criterion;
+use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 
 use rand;
 use rand::Rng;
@@ -16,11 +13,10 @@ use bulletproofs::{BulletproofGens, PedersenGens};
 static AGGREGATION_SIZES: [usize; 6] = [1, 2, 4, 8, 16, 32];
 
 fn create_aggregated_rangeproof_helper(n: usize, c: &mut Criterion) {
-    let label = format!("Aggregated {}-bit rangeproof creation", n);
+    let mut group = c.benchmark_group(format!("Aggregated {}-bit rangeproof creation", n));
 
-    c.bench_function_over_inputs(
-        &label,
-        move |b, &&m| {
+    for m in AGGREGATION_SIZES {
+        group.bench_with_input(BenchmarkId::from_parameter(n), &m, |b, &m| {
             let pc_gens = PedersenGens::default();
             let bp_gens = BulletproofGens::new(n, m);
             let mut rng = rand::rng();
@@ -42,9 +38,8 @@ fn create_aggregated_rangeproof_helper(n: usize, c: &mut Criterion) {
                     n,
                 )
             })
-        },
-        &AGGREGATION_SIZES,
-    );
+        });
+    }
 }
 
 fn create_aggregated_rangeproof_n_8(c: &mut Criterion) {
@@ -74,11 +69,10 @@ criterion_group! {
 }
 
 fn verify_aggregated_rangeproof_helper(n: usize, c: &mut Criterion) {
-    let label = format!("Aggregated {}-bit rangeproof verification", n);
+   let mut group = c.benchmark_group(format!("Aggregated {}-bit rangeproof verification", n));
 
-    c.bench_function_over_inputs(
-        &label,
-        move |b, &&m| {
+    for m in AGGREGATION_SIZES {
+        group.bench_with_input(BenchmarkId::from_parameter(n), &m, |b, &m| {
             let pc_gens = PedersenGens::default();
             let bp_gens = BulletproofGens::new(n, m);
             let mut rng = rand::rng();
@@ -104,9 +98,8 @@ fn verify_aggregated_rangeproof_helper(n: usize, c: &mut Criterion) {
 
                 proof.verify_multiple(&bp_gens, &pc_gens, &mut transcript, &value_commitments, n)
             });
-        },
-        &AGGREGATION_SIZES,
-    );
+        });
+    }
 }
 
 fn verify_aggregated_rangeproof_n_8(c: &mut Criterion) {